feat: spades - show scores in bid dialog
