+++ src/components/games/spades/ScoreTable.tsx
@@ -1,4 +1,5 @@
 import { ScoreRow } from "../../../jotai/spades-atom";
+import { getScoreText } from "./helpers/getScoreText";
 import {
   Table,
   TableBody,
@@ -11,16 +12,6 @@ interface ScoreTableProps {
   initials: string;
   data: ScoreRow[];
 }
-
-const getScoreText = (score?: number, comp?: number) => {
-  if (score === undefined || comp === undefined || score === 0) {
-    return "";
-  } else if (score > 0) {
-    return `${score >= 100 && score >= comp ? "🎉 " : ""}${score}`;
-  } else {
-    return `${score}0 + `;
-  }
-};
 
 const ScoreTable = ({ initials, data }: ScoreTableProps) => (
   <Table aria-label="Bid Table">
