+++ src/components/games/spades/helpers/getChipColor.ts
@@ -1,0 +1,6 @@
+export const getChipColor = (a: number, b: number) => {
+  if (a === b) {
+    return "default";
+  }
+  return a > b ? "success" : "error";
+};
